refactor: move health route, user schema; remove empty legacy dirs
- api/routes/health.py → health/routes.py
- schemas/user.py → auth/schemas.py (only contains UserResponse for /api/me)
- Delete empty api/, schemas/, services/ directories
- Update main.py to use all domain-based router imports
